# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TITAN API",
    "summary": "Self-hosted AI assistant, planner and tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices/pair": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Pair a device with username and password",
        "operationId": "pair_device_api_v1_devices_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or locked account",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "The authenticated user",
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "The caller's paired devices",
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceOut"
                  },
                  "type": "array",
                  "title": "Response List Devices Api V1 Devices Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Revoke a device",
        "description": "Users revoke their own devices; the owner can revoke any device.",
        "operationId": "revoke_device_api_v1_devices__device_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "All accounts (owner only)",
        "operationId": "list_users_api_v1_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "type": "array",
                  "title": "Response List Users Api V1 Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create a member account (owner only)",
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/session": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Sign a browser in",
        "description": "Creates a `web` device and sets its token in the `__Host-TSID` cookie; the token is never in the body. Needs the `X-Titan-Request: 1` header.",
        "operationId": "sign_in_api_v1_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or locked account",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Cross-site request",
            "content": {
              "application/problem+json": {}
            }
          },
          "429": {
            "description": "Too many sign-in attempts from this address",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Sign out: revoke the calling device and clear the cookie",
        "operationId": "sign_out_api_v1_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/session/confirm": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Confirm the password before a sensitive change",
        "description": "Renews the sign-in time of the calling device, without a new device. Sensitive changes from a browser signed in more than 15 minutes ago answer `403` with the problem type `urn:titan:problem:confirm-password` until this is called.",
        "operationId": "confirm_api_v1_session_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 1024,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "ConfirmRequest"
      },
      "CreateUserRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "title": "Password"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "CreateUserRequest"
      },
...
      "DeviceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "current": {
            "type": "boolean",
            "title": "Current",
            "description": "True for the device making this request"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "platform",
          "created_at",
          "last_seen_at",
          "revoked_at",
          "current"
        ],
        "title": "DeviceOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "title": "Password"
          },
          "device_name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Device Name",
            "examples": [
              "Pixel 9"
            ]
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "device_name",
          "platform"
        ],
        "title": "PairRequest"
      },
      "PairResponse": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Shown once. Send as 'Authorization: Bearer <token>'."
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "token",
          "user"
        ],
        "title": "PairResponse"
      },
...
      "Platform": {
        "type": "string",
        "enum": [
          "android",
          "web",
          "cli",
          "other"
        ],
        "title": "Platform"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "member"
        ],
        "title": "Role"
      },
      "SessionOut": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "user"
        ],
        "title": "SessionOut"
      },
      "SignInRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "SignInRequest"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Device token from POST /devices/pair",
        "scheme": "bearer"
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "description": "Set by POST /session; browsers only",
        "in": "cookie",
        "name": "__Host-TSID"
      }
    }
  }
}